requirements.txt
aiofiles==0.8.0
anyio==4.4.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
commonmark==0.9.1
exceptiongroup==1.2.2
fdsauth==0.0.4
Flask==3.0.3
Flask-Cors==5.0.0
geojson==2.5.0
geopandas==1.0.1
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.8
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
multipledispatch==0.6.0
networkx==2.8.8
ngsildclient==0.5.2
numpy==2.1.1
packaging==24.1
pandas==2.2.2
Pygments==2.18.0
pyogrio==0.9.0
pyproj==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
rfc3986==1.5.0
rich==12.6.0
scalpl==0.4.2
shapely==2.0.6
six==1.16.0
sniffio==1.3.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.4
